Minor improvements, final.exe1

+++ Assets/Scripts/Player/HouseHealth.cs
@@ -4,8 +4,8 @@ using UnityEngine;
 
 public class HouseHealth : MonoBehaviour {
 
-    public int maxHealth = 100;
-    public int currentHealth;
+    public float maxHealth = 100;
+    public float currentHealth;
     public bool isDead;
     public bool isDamaged;
